requirements.txt
av
faiss-gpu
scipy
python-dotenv
sounddevice
soundfile
torch